% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunction.R
\name{rdviz}
\alias{rdviz}
\title{Visualize a Regression Discontinuity}
\usage{
rdviz(x, y, c = 0, p = 1, kernel = "tri", xlab = "X", ylab = "Y",
  binnedPoints = T)
}
\arguments{
\item{x}{The running variable}

\item{y}{The dependent variable}

\item{c}{The RD cutoff}

\item{p}{The polynomial order (NOTE: currently only works for p=1 and p=2)}

\item{kernel}{The kernel ('triangular' or 'uniform')}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{binnedPoints}{A boolean. Display the binned scatter plot or no?}
}
\value{
A ggplot object
}
\description{
Visualize a Regression Discontinuity
}
\examples{
rdviz(x,y)
}
